After quite a while, we have another big SVN update now. Most of the GUI code has been reorganized, and we have a working buffer management now. Also, we switched the main window to the new layout, which is probably here to stay; of course it will be enhanced quite a bit in the future :)
Backlogs will be created and displayed whenever the GUI reconnects to the core. We limit
the shown lines to 100 to not use too much time for rendering. This will be configurable in the
future.

There is probably a lot more changes that happened over the past few days which I can't remember,
so you'll have to just see for yourself.