Finished the fusion of the different main_* files, so now we have only one #ifdef-ridden main.cpp.
Thought it might be a good idea to enable gcc warnings (-Wall -Wextra -ansi), so I did. While fixing
most of the resulting warnings, I also found a couple of bugs. So this was a Good Thing[tm], obviously.

Don't even think of enabling -pedantic as well - this fails horribly somewhere in Qt.