OK Folks, my first commit after quite a while, and while Quassel looks the same as before, a lot of changed under the hood. This includes
* Parts of the new settings system, includin session data which is gonna replace Global
* Multi-user core
* Complete separation of core and client even in th monolithic client


Basically now most of the multi-user core and core connection code is in place, as well as
most of the skeleton for the authentification. Since parts of the UI, disconnecting from a core
and connecting to a remote core don't work yet, I have most of it disabled or hidden. Things
will fall back in place piece by piece, without requiring too many changes to the source in
the future I hope.

Oh yeah, due to the new settings stuff, you will have to enter identity, network and other data anew.
Maybe this is a good opportunity to wipe out your .config...