First check-in of the Qtopia build system. See qtopia-build/README.
Yes, this compiles and links an example application (that includes all quassel parts
intended for Qtopia). Nope, it does not display more than a button, since the main() is
still the one from the Qtopia example app.
And yes, this has hardly been tested and will probably be redesigned at some point. But it's still
a proof of concept :)