I CAN HAZ A STYLE ENGINES!?
Say hello to the new module "uisupport", which shall contain code that is shared
across UI implementations (and may depend on QtGui). For now, this is our brand-new
shiny style engine that replaces the dirty hack that was style.cpp/h. UI styles now derive
from UiStyle and set the text formats as they wish through an easy API.
ChatLine has been switched to use QtUiStyle, but since ChatLine might be removed soon, I
did not bother to fully implement this, which in particular means that text selections are
no longer displayed.