Allow user to provide a key/certificate pair for outgoing IRC connections.
OFTC supports a way to identify clients based on the fingerprint of
the SSL certificated used to initialise the connection.

This patch allows to make use of the CertFP identification by creating
a ~/.quassel/quasselClientCert.pem key/certificate file, which will be
used for all the outgoing SSL connections to IRC servers.

At the moment the message sent by the server at the connection is not
displayed by quassel, so you'll have to find the certificate's
fingerprint through OpenSSL (or any other method).

Signed-off-by: Manuel Nickschas <sputnick@quassel-irc.org>