Introduce AbstractMessageProcessor and its specialization QtUiMessageProcessor.
Its task is to create ChatLineModelItems from Messages, which currently happens in Client.
As this is too slow, we will once again process incoming messages, in particular backlog
packages, asynchronously. Externalizing this and abstracting it away from Client will
facilite this. Also the specialization in QtUi will allow us to use GUI features, which will
be particularly handy for determining if we can do font rendering outside the GUI thread.