Add Icon and IconLoader
The former is a very thin wrapper around QIcon that will use the latter to locate an icon
in the current theme (harcoded to Oxygen for now), either in the file system or in a resource.
This will allow us to make icon resources optional, and also to seamlessly integrate with KDE's
icon theming later.
Note that we don't provide proper theming support now, as IconLoader is _very_ rudimentary for
now and assumes Oxygen to be present.