Sanitizing clipboard handling
Quassel now respects the difference between the clipboard and the X11 selection, and won't
put stuff into the clipboard without a) Ctrl+C or b) the new shiny context menu in ChatView.
This means that just selecting stuff won't overwrite your clipboard on Win and Mac, and on X11
only if your system is configured that way. Oh, and we have a context menu now, and Ctrl+C works ;-)