Introduce NetworkModelActionProvider
This is a central place for requesting a list of actions for any given (set of) NetworkModel items.
This allows us to create context menus easily in various places without duplicating code. Also,
NetworkModelActionProvider handles whatever it can internally, so there is no need to handle most
context menu actions elsewhere. This is not fully done yet; so far, we only provide BufferItem and NetworkItem
actions.