Introduce automagic child widget handling in SettingsPage
This tackles the most annoying thing about settingspages: the need for tracking
the changes of every child widget, comparing it with the stored value, and updating
the changedState for the whole page. Now standard widgets only need to have two
dynamic properties set in Qt Designer and all that is handled automatically for them,
including load/save/defaults. See settingspage.h for more information about how to do that.