Properly focus BufferView on first click
Now you don't need to click in a BufferView twice anymore to gain keyboard focus,
for example for hiding a buffer per pressing backspace.
Side-effect is that the input line loses focus now when clicking in the view (obviously),
but the next commmit focuses it automatically on key press, so it should be smooth that way.