Introduce presets for adding networks
We now allow providing a networks.ini that contains a predefined list of networks/servers.
When adding a network, one can choose a preset from that list to have server properties
filled in automatically. This should make it much easier for users to add networks.

For adding a custom network, we allow entering the most important data right in the dialog,
to avoid useless clicks.