Add ToolBarActionProvider as another specialization of NetworkModelController
As toolbar actions need to be handled very differently from context menu actions, we use
another provider for them. This is supposed to keep track of MainWin's state and enable actions
as appropriate. In this first version, most actions are still not functional, as the whole
state-keeping thing hasn't been implemented yet ;-)