Move font settings to general appearance settingspage
This retires our old fonts settingspage for now, since most of what was on
there didn't work anyway. You can now choose fonts for the ChatView, the input line
and (this is brandnew!1!!) the buffer views and nick list.

Also added tooltips for that settingspage and changed the wording a bit.