Add -DLINK_EXTRA to the buildsys to allow for specifying additional libs to link
This is intended for adding missing libs for static builds, as deps are not always picked
up correctly. As it is highly system- and configuration-specific, we can't hardcode those libs
in the buildsys.

For example, to build a static core with postgres support, I need:
-DLINK_EXTRA=pq,crypt

Also fix a warning for WIN32.