Rework mergedFormats() to handle pre-set formats
This allows us to seed the format cache with information from a stylesheet.
mergedFormats() now handles sensible combinations of pre-set values in order
to construct the requested format. For example, if the stylesheet sets a format
for the sender column in a highlighted message, this is now used to construct
derivations from that rather than using the hardcoded stock formats.