Introducing QssParser
This parses our extensions to Qt stylesheets. So far, we can modify the
application palette as follows:

Palette {
  base: white;
  text: rgb(128, 128, 255);
  window-text: #12ffaa33;
  ...
}

The roles and brush/color definitions follow the Qt stylesheet syntax; note that
we allow for styling all of the palette roles (including things like tooltip-base,
which are not included in Qt's list of stylable roles).