Introduce a notification type and add extra notifications for KNotify
We have 4 types of notifications now: for queries and for highlights,
each with and without Quassel having focus. If using KDE support, each
can now be configured separately (see #571).

The non-KDE notification backends keep the old behavior (i.e. only notify if not
focused) for now, until we get a configuration UI for this.