Properly save and restore the mainwindow state
We now save size, position, maximize, minimize and hidden-to-tray on exit
and restore it at next start. Since Qt's restoreState() has some issues, I had
to workaround things a bit, so please test on your system and report if there are
still problems with the current implementation!

Fixes #677, fixes #452 (since the state is remembered, so just quit Quassel using the
tray context menu while it's hidden).