Handle all message types properly in UiStyle; eliminate msgtype format codes
Turned out we didn't handle all message types (as in Message::Type) in
UiStyle, e.g. the daychange message was treated like an error. Possibly this
also fixes the "daychange msg not translated" bug, but I haven't tested yet.

Also, styleString() now takes a base message format as optional parameter. This
eliminates the need to add the msgtype as one of the %Dx format codes in every message.
Accordingly, I've removed those format codes and added a mapping for Message::Type -> UiStyle::FormatType
instead that is used to set the initial format for a string to be styled.

Sorry translators, this touches most of the strings in UiStyle, so those will have to be updated.