Make sender-hash based styling ("colored nicks") work again
Plain messages now obtain their format depending on the sender hash.
This can be used to implement colored nicks, but with our new style engine,
we can now solve several problems like using other colors for highlighted messages.

Note that the number of hashed is currently hard-coded to 16; waiting for a nice way
to make this number configurable in a stylesheet.