Parse mIRC colors in stylesheets
Use a condition like [fg-color="03"] or [bg-color="05"] to specify formats for
mIRC colors. We use this more verbose format (rather than adding extra attributes
to ChatLine instead, like "color-03: red;") to allow for maximum flexibility in
styling. The idea is to be able to specify formats e.g. for a particular combination
of foreground and background.