Revamp ColorButton
Rather than requiring parent widgets using this to connect all sorts of signals and show the
color selection dialog themselves, move this into ColorButton itself, making it a
self-sufficient widget. Also, we're now using the Qt property system, and emit a
colorChanged() signal.

Oh, and we inherit from QToolButton now, which seems to be a saner choice on MacOSX and also
allows for nicer layouts on all platforms. Yes, we'll try and improve the looks later on, right
now it's just a crude QPixmap overlaying the button.