Make the Settings caches flat structures
As settingspages tend to access values by absolute path (rather than by group and key),
we should use a flat hash rather than a nested tree structure, such that caching and
in particular the notifiers work with both the group/key combination and the absolute key.
This should also be slightly more efficient.