Introduce the new ChatViewSettingsPage
This is supposed to group the ChatView-related settings in once place, including fonts
and colors. Oh, and you now also can configure the timestamp format :)

We've reduced the number of configurable colors, simply because more fancy stuff can and should
be done via stylesheets.
Note that the style settings, while stored, are still unbeknownst to the style
engine until we start actually generating stylesheets.