Update ChatView properly on style changes
Since we have the cached wraplists back for the moment, we need to invalidate the caches
for all ChatLineModelItems before relayouting the scene. So instead of having each ChatView trigger
the layout change for his scene, we now inform the ChatLineModel itself, which invalidates its
items' caches and then emits a dataChanged() signal that is caught by the scenes.