Remove label from FontSelector
Earlier the FontSelector had its own label, which in retrospect was a stupid idea.
Much better to just consider the demo widget and choose button parts of the FontSelector
widget and have settingspages supply their own description label - or checkbox.

This makes it much easier to have a font disableable, without introducing the need for
multi-property autowidgets.