s/Buffer/Chat/g
This removes the mention of "Buffer" from the UI, as it was confuzzling users and making our Usability Queen
unhappy. It has been replaced by "Chat", which might not always be as precise, but should make more sense for users.
A "Buffer View" is now a "Chat List", and in some cases, "Window" has been used (e.g. the "Status Buffer" is now
a "Status Window" - which, again, isn't really the correct term, but probably makes more sense to people not used
to cli-based chat clients.

Obviously, within the code and the DB everything is still a Buffer (and it'll stay that way).