Add ResizingStackedWidget
This extends QStackedWidget by adding a dynamic sizeHint that only depends on the
currently selected page. Geometry is updated on page change. Use this widget if you want
a stacked container that dynamically adapts to the size of the visible page only rather than
caring for all the pages. Note that pages should have QSizePolicy::Ignore for this to work properly.