Introducing StyledLabel
This widget evolved from TopicLabel and provides a QFrame that contains a mIRC-formatted string
(e.g. it accepts mIRC's format codes in the input). Rather than the old drag-to-scroll behavior,
this widget provides a tooltip. It also uses proper QTextLayouts instead of assembling text snippets
manually.

Note that URLs are not yet clickable. This will be added later.