Try workarounding bug #663 ("Teh Systray Heisenbug")
Setting uniform row heights for BufferView seems to keep Quassel from crashing. This is not
a long-term solution and has some sideeffects (don't use different font sizes for network and buffer
items when playing with QSS!), but if it helps people to start Quassel, fine with me.

Please report if it fixes the crashes for you!