Make installing Oxygen icons optional
The new CMake option -DWITH_OXYGEN=(ON|OFF|AUTO) now controls whether the Oxygen icons
Quassel uses are to be installed. If AUTO (the default) is selected, icons will not be installed
iff KDE > 4.3.0 is present and enabled, because we can then expect them to be present in the
system theme.

Also updated the importer script and docs accordingly.