Untie the marker line from lastSeenMsg
We now handle the marker line position independently from lastSeenMsg in the backend. This will
allow for setting the marker line manually in the future, for people who don't want the automatic
mark-as-read thingy. Currently, it's still set on buffer change, and the behavior should be the same
as before.

Note that this needs an updated core to work correctly, and will be disabled (= forced auto) else.