Make sure that text lines aren't squeezed together in new Qt versions
In recent Qt 4.6 versions, there is a commit[1] that changes behavior wrt to lineSpacing() and
friends. Since (at least on my box) the font leading (the distance between text lines) seems to be
always -1 (for reasons as of yet unbeknownst to me), using lineSpacing() lets text lines in ChatView
overlap. This workarounds this with using at least height().

[1] http://qt.gitorious.org/qt/qt/commit/04d18b38c38c5ff623b30366ea08d56128b9b7d0