Merged duplicate msgid and translations with msguniq
In order for gettext tools such as msgmerge and msgfmt to work with PO
files, these must have a valid format. Therefore duplicate msgid must be
merged into single ones, which is what has been done to the current
translations. The following command has been used (using Italian as an
example):

  msguniq --use-first quassel_it.po > it.po && rm quassel_it.po

Note that some of the PO header files had to be fixed manually prior to
executing the command, so that they had the right format.

Additionally, the file names have been renamed to ll.po (where ll is the
language code). While this is just a minor detail, I believe it makes
sense to use the same naming convention as gettext-based OSS projects.