Don't explicitly check for OpenSSL anymore, but rely on Qt
As we're not compiling against or linking to OpenSSL, but only use Qt's SSL-related
classes, it makes no sense to check for OpenSSL support ourselves. Instead, we now check
if Qt has been built with SSL support and enable our own SSL stuff in exactly this case.

This relies on Qt caring about its deps itself, though, as it already is done for the
other Qt modules.