Properly clean up after disconnecting
Mainly this fixes a crash on reconnect, because ToolBarActionProvider didn't clean
up its state after disconnecting. Now the relevant classes have slots for connecting
and disconnecting that can be reimplemented if necessary.

A better solution is probably polling Client's state rather than using slots, but for
now this should be ok.