Reset the BufferViewFilter after setting a new config
I'm not sure why, but somehow BufferView crashes in rare occasions because of invalid modelindexes
if we don't reset the BufferModel after setting its config. This is even true if we haven't set
a source model before, so an attached View couldn't even get any invalid indexes in the first
place. Anyway, it stops Quassel from crashing here, so I shouldn't complain, I guess.

This hopefully fixes #663, the Heisenbug that has haunted us for a long time now. At least for me.