Stop RFC-violating IRC servers from crashing quassel
Seriously, '%' is not allowed in nicknames. But who cares about RFCs...

Escaping this now in the nickname as well, and additionally prevent the style engine
from hanging itself if it still encounters an unescaped % (even though that can't happen now
anymore). Thanks to ricky26 for the heads up and a patch.

Fixes #946, fixes #947