Close notifications on other connected clients as well
Client now emits a bufferMarkedAsRead() signal that is used for triggering
the closing of notifications. This is synced between clients.

This requires both uptodate clients and core and will spam your log with warnings
if one is too old (but continue to work like it used to). "Too old" means "does not
contain this commit", i.e. a 0.6.0 client will work fine with a git master core and
the other way round, but earlier versions won't.