Improve marker line behavior; allow manual setting (Ctrl+R)
The marker line will now be set to the bottom of the current chatview rather than
to the very bottom of the chatview when it makes sense. So if you didn't scroll all
the way down, and switch channels, you can start off where you left reading.

Also there's now an option to disable the automatic setting of the marker line completely.
Using Ctrl+R, you can manually set the marker line to the bottom of the ChatView.

This requires a recent core (older cores will still work, but only show the old behavior).