Cache ChatLine layout data for visible chatlines
Up to now, we've computed e.g. the text layout for a chatline on the fly. This
happens surprisingly often, for example whenever the mouse is moved. As a result,
moving the mouse (among other things) caused high CPU load.

We now cache such data within the chatitems until the chatline goes out of view, in
which case the cache is cleared in order to not waste space needlessly.

Fixes #954.