Fix keysequence strings for Mac again
Qt actually offers QKeySequence::NativeText to do Mac special casing for toString().
So it makes sense to use that. We now also simulate Qt's behavior for modifier names, that is,
on Mac those will be unicode symbols and no '+' is set between keys. Also the order should be
the commonly used one now (I hope).