Adapt the stylesheets and the stylesheet generator to the new markerline
The markerline now supports a fully styled brush rather than just reading the color
and enforcing a gradient.

In order to get the same look as before, we adapt the stylesheets we ship and also write
out an appropriate gradient definition in settings.qss.

Note that, as always in QSS, the gradient is normalized between (0,0) and (1,1).