Fix wordwrap when using Qt > 4.6.3
Qt changed the behavior of QTextBoundaryFinder in commit e6ac173, to make it more
consistent and to fix a bug in QtWebkit (cf. https://bugs.webkit.org/show_bug.cgi?id=31076).

Unfortunately this means we have to remove our wordwrap workaround for newer versions of Qt,
so we're adding a runtime version check for that. Please note that if you're using a 4.7 snapshot
that's too old, you'll have broken wordwrap now. Upgrade Qt.