Force parsing of PREFIX in RPL_ISUPPORT. Fixes #936
In some cases the automatic parsing of PREFIX was actually
triggered before RPL_ISUPPORT was actually received. In this case
the internal prefix configuration would only contain the default
modes, and not the network specific modes, which would cause
custom modes to be ignored completly.

This patch forces the parsing of PREFIX directly after RPL_ISUPPORT
was received, and the data is available.