Rework message splitting to properly handle encrypted messages
Since cutting off a crypted message produces garbage, and we can't reliably predict
how long the encrypted version will be in relation to the cleartext string, we need
to do some extra trickery in the message splitting code.

As a side effect, the split char now also remains on the previous line, which is
aesthetically way more pleasing than starting a line with a space or punctuation.
trickery for this case. Since