Add support for dbusmenu
If libdbusmenu-qt is present, Quassel should now export the tray menu via D-Bus.
Note that I could not test this as my KDE doesn't seem to care about dbusmenu here...
Note also that icons in the tray menu won't be displayed for now.

This also adds the new IconThemePath() method for the StatusNotifier.