Properly reject invalid handshake data
Currently, if a client attempts to configure a core's storage backend
even though it is already configured, the core will ignore the attempt
silently (e.g. the existing storage backends will simply drop the request,
but the client also won't receive an error message). This can never happen
unless you use an alternative client with a buggy handshake sequence though,
since quasselclient checks if the core is already configured beforehand.

Nonetheless, it makes sense to properly error out from the core side, just in case
we're getting new storage backends in the future that don't properly check this,
or alternative clients which might be confused by not receiving an error when they should.

Thanks to JodaZ for pointing this out.