Add a timestamp to events
This is not set automatically for now, as I'm not sure if it's worth the overhead
of querying the system clock for every event we create.

The timestamp is set on reception of data on the socket (i.e. a bit earlier in the chain
than in the old message handling) and carried over into IrcEvents by the IrcParser.
EventStringifier will carry over the timestamp to MessageEvents.