Implement the ChatLine element in C++
Now that I get my brain wrapped around QML a bit, let's take things
more seriously. Turns out, while Text elements seem to be efficient enough,
TextEdit elements (needed for selection and mousehandling) are not, so resizing
is sluggish.

Thus, instead of using any of the stock elements, we're now implementing this ourselves
in C++, and will probably do much the same as we did for the QGV-based ChatView (minus
some of the hackery, if we're lucky). This means layouting and drawing the text ourselves.
On the upside, we can just keep using the old styleengine for now...

Here's hoping that things will turn out to be more efficient anyway, because ListView seems
to be creating and discarding its items as needed; so there shouldn't be a scene with thousands
of items involved to manage...