Bring Quassel to front on tray icon click
When the quassel window is below other windows it would be nice to have it raised on top of
the others when clicking on the system tray icon. Now it is just hidden and you have to click
on it once more to get it to the top.

Fixes #1135. Thanks to parapente for the patch!