Also fix the icon naming in hicolor.qrc
One remark concerning this naming change: in order to fix the new monochrome tray icon stuff
in >= KDE 4.8, not only the icon naming spec fix in Quassel is needed, but also the desktop themes
need to be fixed (so their quassel.svgz contains the changed filenames). This is in upstream kde-runtime,
but at this time not backported to KDE 4.8.

Thanks to Kai Uwe Broulik for pointing me to the reason for the broken icons and committing the
needed changes to kde-runtime.