Fix output for a bunch of numeric server replies
First of all, takeFirst() does not work on a const ref *cough*. And second, we should
not reorder the param in the message like we did before purely for prettyness;
some servers seem to send non-standard strings that then look awkward. Instead,
let's just add a colon after the first param.