Add support for Qt5 in the build system
Support for building against Qt5 can now be enabled by passing -DWITH_QT5=1 to CMake.
Note that your Qt5 installation (the qmake location) needs to be in your PATH, unless
it's already in one of the standard system locations.

Needless to say that Qt5 support in Quassel is highly experimental, though surprisingly
it seems to work mostly fine already. A README.Qt5 with known issues is forthcoming.

As a side effect, I've cleaned up the build system a bit and got rid of QT_USE_FILE and the related
fugly hacks for Qt4 (so that the -I and -D options now actually match the modules we're building
against).