Provide (de)serialization for all event types
This allows to (de)serialize events to/from QVariantMap (using only primitive
types), a prerequisite for sending events over the wire.

Use the following functions:

void EventManager::createEvent(const QVariantMap &)
QVariantMap Event::toVariantMap()

Note that there is only rudimentary plausibility checking, so for now we assume
that the QVariantMaps are created by Quassel. Before we allow these to go on the wire,
we need to make sure that this can't be tinkered with in annoying ways.