Relicense SystemTray and LegacySystemTray under LGPL 2+
This was requested so that it could be made useful together with
StatusNotifierItem (which came from KDE and already is LGPL'd).

Here's hoping that it will result in a neat little library that programs
can use to have a systemtray icon that sucks less than the stock one.

This code was written by yours truly.