Fix building with Qt 4.7
For reasons unknown, qPrintable() doesn't work with QByteArray in Qt versions
prior to 4.8 (the confusing thing is why it works now, in fact). In any case,
it probably makes sense to print a binary blob using normal QDebug output for
QByteArrays anyway, rather than trying to convert it to a string.