Implement issue #1219: Update the marker line when quassel loses focus
Solves issue #1219 on bugs.quassel-irc.org.

We now listen for QEvent::WindowDeactivate in the main window. This
should pickup on window closing/minimizing/losing focus. We also change
view->setMarkerLine() to Client::setMarkerLine which will send an event
to the syncher, and eventually arrives at ChatView::markerLineSet, which
moves the marker line for us.